reduce buffer sizes for test suite since it forces lots of small flushes.  reduces NameSortTest from ~30s back to < 10s. patch by jbellis for CASSANDRA-339
git-svn-id: https://svn.apache.org/repos/asf/incubator/cassandra/trunk@802276 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/cassandra/io/SSTableWriter.java
@@ -27,8 +27,8 @@ public class SSTableWriter extends SSTable
     public SSTableWriter(String filename, int keyCount, IPartitioner partitioner) throws IOException
     {
         super(filename, partitioner);
-        dataFile = new BufferedRandomAccessFile(path, "rw", DatabaseDescriptor.getFlushDataBufferSizeInMB() * 1024 * 1024);
-        indexFile = new BufferedRandomAccessFile(indexFilename(), "rw", DatabaseDescriptor.getFlushIndexBufferSizeInMB() * 1024 * 1024);
+        dataFile = new BufferedRandomAccessFile(path, "rw", (int)(DatabaseDescriptor.getFlushDataBufferSizeInMB() * 1024 * 1024));
+        indexFile = new BufferedRandomAccessFile(indexFilename(), "rw", (int)(DatabaseDescriptor.getFlushIndexBufferSizeInMB() * 1024 * 1024));
         bf = new BloomFilter(keyCount, 15);
     }
